Tests/LibWeb/Text/expected/selection-scroll-into-view.txt

vertical scrollTop: 80
horizontal scrollLeft: 603.171875
